requirements.txt
numpy
matplotlib
pillow
opencv-python
ipykernel